Changes to Clock/Daytime Modulator  and addition of Weather_Timer
I changed the Clock/Daytime Modulator script to have longer nights that can be shortened or extended depends on how we feel. I added the season cycle to the clock script. And I also added the weather conditions using enums for now in the timer script . Cleared some bugs in my code for the scripts , created a pickup folder